Initial commit
This is the initial code for « Capteur de vide multilignes ».
Read the temperature, light intensity and 4 vacuum channesl and publish
theses measurements along with battery voltage, percentage of charge,
Signal strength and quality, count of bytes transferred and the time
the Electron is ON.

It awaken every 2 minutes for test purpose. This will be changed to 20
minutes later. A button on the system also permit to awake it at will.

The measurements routines were validated on a Photon first.